extract op stack analysis helpers
